Rename config to VTConfig
